backend/accounting/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
pydantic==2.9.2
python-multipart==0.0.20
httpx==0.27.2
anthropic==0.39.0

# Auth + rate limit + logging
PyJWT==2.10.0
slowapi==0.1.9

# 測試(選配)
pytest==8.3.3
mongomock==4.2.0.post1
